chess-db: drop a redundant u64 cast (clippy 1.96 unnecessary_cast)
CI's stable clippy (1.96) flags a u64->u64 cast that the older local
toolchain (1.94) didn't. total_kb is already u64.

+++ chess-db/src/db/mod.rs
@@ -12,7 +12,7 @@ fn memory_limit_str() -> String {
     if total_kb == 0 {
         return "8GiB".to_string();
     }
-    let limit_gib = (total_kb as u64 * 80 / 100) / (1024 * 1024);
+    let limit_gib = (total_kb * 80 / 100) / (1024 * 1024);
     format!("{}GiB", limit_gib.max(1))
 }
